Add support for listing item types other than dictionaries
Although the /content-codes endpoint reports that the mlit code is of
the list type (12), this is not always the case. For example, the
daap.browseartistlisting type uses mlit fields with string data. To
handle these messages a parser needs to be aware of the parent context
as well as the content-code to determine what type of data is expected
for the mlit code.

To accomplish this the dmap_field structure is modified to include an
additional DMAP_TYPE field storing the expected type of the listing item
children. Parsing is moved to a dmap_parse_internal() function that
receives a pointer to the parent dmap_field, if any. Upon encountering
an mlit field the parent field's list_item_type is used to parse the
listing item. If a listing item type is not specified a list/dictionary
is assumed.