Do not pass negative values to character classification functions
This fixes a debug assertion on Windows. The Windows implementation of
the classification functions converts the input value to an unsigned
integer and verifies that it is representable as an unsigned char as
required by the standard. If the sign bit is set on the input value the
value will be outside of the range of an unsigned char and trigger the
assertion.